domains/discord/ingestion/pyproject.toml
[project]
name = "discord-etl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "orjson>=3.10",
  "click>=8.1",
  "pyyaml>=6.0",
  "pydantic>=2.9",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.8",
  "pytest-httpx>=0.35",
  "types-pyyaml>=6.0",
]

[project.scripts]
discord-etl = "discord_etl.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
